requirements.txt
ipyKernel 
numpy
pandas
openpyxl
python-dotenv
matplotlib
seaborn
